refactor(sidebar): split sections and rename sidebar modules
