+++ src/components/layout/sidebar-modules/FavoriteItem.tsx
