+++ src/components/layout/sidebar-modules/index.ts
@@ -1,0 +1,18 @@
+export { FavoriteItem } from "./FavoriteItem";
+export { FavoritesSection } from "./FavoritesSection";
+export { NavSectionCollapsible } from "./NavSectionCollapsible";
+export { NamespaceSection } from "./NamespaceSection";
+export { PortForwardsSection } from "./PortForwardsSection";
+export { QuickAccessSection } from "./QuickAccessSection";
+export { RecentSection } from "./RecentSection";
+export { ReconnectingTimer } from "./ReconnectingTimer";
+export { implementedViews } from "./constants";
+export { useNavigationSections } from "./navigation";
+export { useSidebarUiState } from "./useSidebarUiState";
+export type {
+  NavSection,
+  ResourceType,
+  SidebarProps,
+  SidebarUiState,
+  SidebarUiStateHook,
+} from "./types";
