--- src/components/layout/sidebar-parts/index.ts
@@ -1,6 +1,0 @@
-export { FavoriteItem } from "./FavoriteItem";
-export { NavSectionCollapsible } from "./NavSectionCollapsible";
-export { ReconnectingTimer } from "./ReconnectingTimer";
-export { implementedViews, readSidebarUiState, saveSidebarUiState } from "./constants";
-export { useNavigationSections } from "./navigation";
-export type { NavSection, ResourceType, SidebarProps, SidebarUiState } from "./types";
